software/requirements.txt
argcomplete
keyring
keystoneauth1
keystonemiddleware
lxml
oslo.config
oslo.policy
oslo.serialization
netaddr
pecan
psutil
psycopg2-binary
pycryptodomex
PyGObject
requests_toolbelt
sh
WebOb
WSME>=0.5b2
